+++ docs/metro/js/metro.js
@@ -170,12 +170,18 @@ var Metro = {
         change: 'change.metro',
         cut: 'cut.metro',
         paste: 'paste.metro',
-        drop: 'drop.metro',
         scroll: 'scroll.metro',
         scrollStart: 'scrollstart.metro',
         scrollStop: 'scrollstop.metro',
         mousewheel: 'mousewheel.metro',
-        inputchange: "change.metro input.metro propertychange.metro cut.metro paste.metro copy.metro"
+        inputchange: "change.metro input.metro propertychange.metro cut.metro paste.metro copy.metro",
+        dragstart: "dragstart.metro",
+        dragend: "dragend.metro",
+        dragenter: "dragenter.metro",
+        dragover: "dragover.metro",
+        dragleave: "dragleave.metro",
+        drop: 'drop.metro',
+        drag: 'drag.metro'
     },
 
     keyCode: {
@@ -9762,7 +9768,10 @@ var File = {
         return this;
     },
     options: {
+        mode: "input",
         buttonTitle: "Choose file(s)",
+        dropTitle: "<strong>Choose a file</strong> or drop it here",
+        dropIcon: "<span class='default-icon-upload'></span>",
         prepend: "",
         clsComponent: "",
         clsPrepend: "",
@@ -9796,9 +9805,9 @@ var File = {
         var that = this, element = this.element, o = this.options;
         var prev = element.prev();
         var parent = element.parent();
-        var container = $("<div>").addClass("file " + element[0].className).addClass(o.clsComponent);
+        var container = $("<label>").addClass((o.mode === "input" ? " file " : " drop-zone ") + element[0].className).addClass(o.clsComponent);
         var caption = $("<span>").addClass("caption").addClass(o.clsCaption);
-        var button;
+        var icon, button;
 
         if (prev.length === 0) {
             parent.prepend(container);
@@ -9807,22 +9816,28 @@ var File = {
         }
 
         element.appendTo(container);
-        caption.insertBefore(element);
 
-        button = $("<button>").addClass("button").attr("tabindex", -1).attr("type", "button").html(o.buttonTitle);
-        button.appendTo(container);
-        button.addClass(o.clsButton);
+        if (o.mode === "input") {
+            caption.insertBefore(element);
 
-        if (element.attr('dir') === 'rtl' ) {
-            container.addClass("rtl");
+            button = $("<button>").addClass("button").attr("tabindex", -1).attr("type", "button").html(o.buttonTitle);
+            button.appendTo(container);
+            button.addClass(o.clsButton);
+
+            if (element.attr('dir') === 'rtl' ) {
+                container.addClass("rtl");
+            }
+
+            if (o.prepend !== "") {
+                var prepend = $("<div>").html(o.prepend);
+                prepend.addClass("prepend").addClass(o.clsPrepend).appendTo(container);
+            }
+        } else {
+            icon = $(o.dropIcon).addClass("icon").appendTo(container);
+            caption.html(o.dropTitle).insertAfter(icon);
         }
 
         element[0].className = '';
-
-        if (o.prepend !== "") {
-            var prepend = $("<div>").html(o.prepend);
-            prepend.addClass("prepend").addClass(o.clsPrepend).appendTo(container);
-        }
 
         if (o.copyInlineStyles === true) {
             for (var i = 0, l = element[0].style.length; i < l; i++) {
@@ -9839,11 +9854,13 @@ var File = {
 
     _createEvents: function(){
         var element = this.element, o = this.options;
-        var parent = element.parent();
-        var caption = parent.find(".caption");
-        parent.on(Metro.events.click, "button, .caption", function(){
+        var container = element.closest("label");
+        var caption = container.find(".caption");
+
+        container.on(Metro.events.click, "button", function(){
             element.trigger("click");
         });
+
         element.on(Metro.events.change, function(){
             var fi = this;
             var file_names = [];
@@ -9856,13 +9873,39 @@ var File = {
                 file_names.push(file.name);
             });
 
-            entry = file_names.join(", ");
+            if (o.mode === "input") {
 
-            caption.html(entry);
-            caption.attr('title', entry);
+                entry = file_names.join(", ");
+
+                caption.html(entry);
+                caption.attr('title', entry);
+            }
 
             Utils.exec(o.onSelect, [fi.files, element], element[0]);
         });
+
+        element.on(Metro.events.focus, function(){container.addClass("focused");});
+        element.on(Metro.events.blur, function(){container.removeClass("focused");});
+
+        if (o.mode !== "input") {
+            container.on('drag dragstart dragend dragover dragenter dragleave drop', function(e){
+                e.preventDefault();
+                e.stopPropagation();
+            });
+
+            container.on('dragenter dragover', function(){
+                container.addClass("drop-on");
+            });
+
+            container.on('dragleave', function(){
+                container.removeClass("drop-on");
+            });
+
+            container.on('drop', function(e){
+                element[0].files = e.originalEvent.dataTransfer.files;
+                container.removeClass("drop-on");
+            });
+        }
     },
 
     disable: function(){
@@ -16854,6 +16897,7 @@ var Table = {
         clsBody: "",
         clsBodyRow: "",
         clsBodyCell: "",
+        clsCellWrapper: "",
 
         clsFooter: "",
         clsFooterRow: "",
@@ -17152,7 +17196,8 @@ var Table = {
                 cls: item_class,
                 colspan: item.attr("colspan"),
                 type: "data",
-                size: Utils.isValue(item.data("size")) ? item.data("size") : ""
+                size: Utils.isValue(item.data("size")) ? item.data("size") : "",
+                show: !item.hasClass("hidden") || (Utils.isValue(item.data("show")) && JSON.parse(item.data("show")) === false)
             };
             that.heads.push(head_item);
         });
@@ -18030,7 +18075,7 @@ var Table = {
                 $.each(cells, function(cell_index){
                     if (o.cellWrapper === true) {
                         td = $("<td>");
-                        $("<div>").addClass("cell-wrapper").html(this).appendTo(td);
+                        $("<div>").addClass("cell-wrapper").addClass(o.clsCellWrapper).html(this).appendTo(td);
                     } else {
                         td = $("<td>").html(this);
                     }